fix: uint转uchar*错误

+++ inc/detect_thread.h
@@ -13,7 +13,7 @@ class DetectThread {
     private:
     static int detect_func(DetectThread *handler);
     Result result_to_uchars(const std::vector<BBox> &detect_result, std::vector<u_char> &send_buffer);
-    u_char uint2chars(double num, std::vector<unsigned char> &cs);
+    u_char uint2chars(uint32_t num, std::vector<unsigned char> &cs);
 
     public:
     int get_queue_id;
